Move bar config into its own file
